project/requirements.txt
streamlit
langchain
langchain-openai
langchain-google-genai
google-generativeai
langchain_huggingface
openai
python-dotenv

